server/pyproject.toml
[project]
name = "customers-talking-server"
version = "0.1.0"
description = "Customer care automation backend — Africa's Talking + Dialogflow CX"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.0",
    "gunicorn>=23.0.0",
    "africastalking>=1.2.7",
    "google-cloud-dialogflow-cx>=1.38.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-flask>=1.3.0",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
